Improve P=3 biclique counting algorithms
- Change triplet sampling from fixed 1M to adaptive 1% of total triplets
- Update Naive algorithm to use exact counting (sampling_ratio=1.0) instead of sampling
- Fix ground truth overflow issues by using long double consistently
- Update test script to use proper parameters for exact counting vs sampling
- Improve algorithm comparison with better sampling strategy

+++ fixed_ground_truth.cpp
@@ -7,7 +7,7 @@
 #include "biclique.h"
 
 // Global variables
-extern unsigned long long real;
+extern long double real;  // Use long double consistently for ground truth
 extern int P___;
 extern int K___;
